move graph transformation state out of csr
