+++ packages/effect/src/internal/graphCsr.ts
@@ -1,5 +1,5 @@
 import type * as Graph from "../Graph.ts"
-import { type GraphImpl, toImpl } from "./graph.ts"
+import { type GraphImpl, isTransforming, toImpl } from "./graph.ts"
 
 /** @internal */
 export interface Adjacency {
@@ -52,7 +52,6 @@ export interface Csr {
 }
 
 const cache = new WeakMap<GraphImpl<any, any, any>, Csr>()
-const transformationDepth = new WeakMap<GraphImpl<any, any, any>, number>()
 
 /** @internal */
 export const getNodeIndex = (csr: Csr, nodeId: Graph.NodeIndex): number | undefined =>
@@ -224,36 +223,11 @@ export const invalidate = <N, E, T extends Graph.Kind>(
 }
 
 /** @internal */
-export const isTransforming = <N, E, T extends Graph.Kind>(
-  graph: Graph.Graph<N, E, T> | Graph.MutableGraph<N, E, T>
-): boolean => transformationDepth.has(toImpl(graph))
-
-/** @internal */
-export const withTransformation = <N, E, T extends Graph.Kind, A>(
-  graph: Graph.MutableGraph<N, E, T>,
-  evaluate: () => A
-): A => {
-  const impl = toImpl(graph)
-  const depth = transformationDepth.get(impl) ?? 0
-  cache.delete(impl)
-  transformationDepth.set(impl, depth + 1)
-  try {
-    return evaluate()
-  } finally {
-    if (depth === 0) {
-      transformationDepth.delete(impl)
-    } else {
-      transformationDepth.set(impl, depth)
-    }
-  }
-}
-
-/** @internal */
 export const get = <N, E, T extends Graph.Kind>(
   graph: Graph.Graph<N, E, T> | Graph.MutableGraph<N, E, T>
 ): Csr => {
   const impl = toImpl(graph)
-  const cacheEnabled = !transformationDepth.has(impl)
+  const cacheEnabled = !isTransforming(graph)
   if (cacheEnabled) {
     const cached = cache.get(impl)
     if (cached !== undefined) {
